+++ src/options/Constants/PageConstants.cs
@@ -17,7 +17,10 @@
         public const string PackageName = Vsix.Name;
         public const string PackageFeatureSet = PackageName + Space + FeatureSet;
         public const string PackageVersion = "Version Number";
+        public const string FeedUrl = "Feed Url";
         public const string OpenLinksInVS = "Open Links in VS";
+        public const string ItemsToDisplay = "Number of items to display";
+        public const string ClearListBeforeRefresh = "Clear list before refresh";
         public const string EnableDeveloperNewsOptions = "Developer News Options";
     }
 }