+++ src/ui/Services/NewsItemCommandService.cs
@@ -12,13 +12,18 @@ namespace DeveloperNews.UI.Services
 
     public class NewsItemCommandService : INewsItemCommandService
     {
-        public ObservableCommandList GetCommands(Action refresh)
+        public ObservableCommandList GetCommands(Action refresh, Action options)
             => new ObservableCommandList
             {
                 new CommandViewModel
                 {
                     Name = "Refresh",
                     Command = new RelayCommand(refresh, true),
+                },
+                new CommandViewModel
+                {
+                    Name = "Options",
+                    Command = new RelayCommand(options, true),
                 }
             };
     }