+++ src/ui/Services/NewsItemDataService.cs
@@ -25,7 +25,7 @@ namespace DeveloperNews.UI.Services
         public NewsItemDataService()
         { }
 
-        public async Task<ObservableCollection<NewsItemViewModel>> GetItemsAsync(string feedUrl, int Count)
+        public async Task<ObservableCollection<NewsItemViewModel>> GetItemsAsync(string feedUrl, int count)
         {
             //https://wp.qmatteoq.com/?p=6486
             //https://blog.qmatteoq.com/the-mvvm-pattern-dependency-injection/
@@ -35,7 +35,9 @@ namespace DeveloperNews.UI.Services
                 var response = await client.GetStringAsync(feedUrl);
                 var document = XDocument.Parse(response);
                 var viewModels = (
-                    from item in document.Descendants(ITEM_ELEMENT_NAME)
+                    from item in document
+                        .Descendants(ITEM_ELEMENT_NAME)
+                        .Take(count)
                     select new NewsItemViewModel
                     {
                         Title = ExtractTitle(item),