+++ src/vsix/Commands/ViewDeveloperNewsOptions.cs
@@ -9,24 +9,23 @@ namespace DeveloperNews.Commands
 {
     using Options.Pages;
 
-
-    internal sealed class DeveloperNewsOptions : DeveloperNewsCommand
+    internal sealed class ViewDeveloperNewsOptions : DeveloperNewsCommand
     {
-        private DeveloperNewsOptions(AsyncPackageBase package)
-            : base(package, PackageIds.DeveloperNewsOptions)
+        private ViewDeveloperNewsOptions(AsyncPackageBase package)
+            : base(package, PackageIds.ViewDeveloperNewsOptions)
         { }
 
         public static async Tasks.Task InstantiateAsync(AsyncPackageBase package)
-            => await InstantiateAsync(new DeveloperNewsOptions(package));
+            => await InstantiateAsync(new ViewDeveloperNewsOptions(package));
 
         protected override bool CanExecute
-          => base.CanExecute && PackageClass.GeneralOptions.EnableDeveloperNewsOptions;
+          => base.CanExecute && GeneralOptions.Instance.EnableDeveloperNewsOptions;
 
         protected override void OnExecute(OleMenuCommand command)
             => ExecuteCommand()
                 .ShowProblem();
 
         private CommandResult ExecuteCommand()
-            => Package?.ShowOptionsPage<GeneralDialogPage>();
+            => Package?.ShowOptionsPage<DialogPageProvider.General>();
     }
 }